uplaod and display file
